.env shortcut

+++ Makefile
@@ -1,5 +1,5 @@
 env:
-	cp .env.example .env
+	if [ ! -f .env ]; then cp .env.example .env; else echo ".env already exists. Skipping."; fi
 
 compose:
 	docker-compose down
